MySignalProcessing

Documentation for MySignalProcessing.

MySignalProcessing.impseq

MySignalProcessing.sigshift

MySignalProcessing.stepseq

MySignalProcessing.impseq — Method

δ(start::Int, stop::Int)

Unit impulse signal This function has the form of :

δ(n) = 1 if n=0
     = 0 if n!=0

parameters

`start:: Int` : start of the impulse sequence n (horizontal axis)
`stop:: Int` : stop of the impulse sequence n (horizontal axis)

returns

s::signal

MySignalProcessing.sigshift — Method

sigshift(s::signal, m::Int)

Function to shift signal s by adding m to each element of s.n It implements y(n) = x(n-n₀)

parameters

`s::signal` : signal to be shifted
`n₀::Int` : shift amount

returns

s'::signal shifted

MySignalProcessing.stepseq — Method

stepseq(start::Int, stop::Int)

Unit Step Sequence This function has the form of :

stepseq(n) = 1 if n ≥ 0
            = 0 if n < 0

parameters

`start:: Int` : start of the step sequence n (horizontal axis)
`stop:: Int` : stop of the step sequence n (horizontal axis)

returns

s::signal