MySignalProcessing

Documentation for MySignalProcessing.

Base.:*

Base.:*

Base.:*

Base.:+

MySignalProcessing._sigscale

MySignalProcessing._sigshift

MySignalProcessing.energy

MySignalProcessing.impseq

MySignalProcessing.realexp

MySignalProcessing.sigadd

MySignalProcessing.sigfold

MySignalProcessing.sigmult

MySignalProcessing.sigrand

MySignalProcessing.sigscale

MySignalProcessing.sigshift

MySignalProcessing.stepseq

Base.:* — Method

*(α::Real, s₁::signal)

scale signal by scalar.

Arguments

`s₁::signal` : signal to be multiplied
`α::Real` : scalar

Returns

`s::signal` : signal s1 * α

Base.:* — Method

*(s₁::signal  , α::Real)

scale signal by scalar.

Arguments

`s₁::signal` : signal to be multiplied
`α::Real` : scalar

Returns

`s::signal` : signal s1 * α

Base.:* — Method

*(s₁::signal  , s₂::signal)

Multiples two signal objects together.

Arguments

`s₁::signal` : signal to be multiplied
`s₂::signal` : signal to be Multiples

Returns

`s::signal` : signal s1 * s2

Base.:+ — Method

Base.+(s₁::signal  , s₂::signal)

Extends Base.+ to add two signal objects together.

Arguments

`s₁::signal` : signal to be added
`s₂::signal` : signal to be added

Returns

`s::signal` : signal s1 + s2

MySignalProcessing._sigscale — Method

_sigscale(s::signal, α::Real)

Each signal sample is multiple by a scalar α

parameters

`s::signal` : signal to be scaled
`α::Real` : scalar

returns

`s::signal` : scaled signal

MySignalProcessing._sigshift — Method

_sigshift(s::signal, m::Int)

Function to shift signal s by adding m to each element of s.n It implements y(n) = x(n-n₀)

parameters

`s::signal` : signal to be shifted
`n₀::Int` : shift amount

returns

s'::signal shifted

MySignalProcessing.energy — Method

energy(s::signal)

Calculate the energy of a signal The energy of a signal is given by the sum of the absolute squared values of the signal samples

parameters

`s::signal` : input signal

returns

`e::Real` : energy of the signal

MySignalProcessing.impseq — Method

δ(start::Int, stop::Int)

Unit impulse signal This function has the form of :

δ(n) = 1 if n=0
     = 0 if n!=0

parameters

`start:: Int` : start of the impulse sequence n (horizontal axis)
`stop:: Int` : stop of the impulse sequence n (horizontal axis)

returns

s::signal

MySignalProcessing.realexp — Method

realexp(a::Real, n₀::Int , nf::Int)

Function to generate a real exponential signal.

Signal of the form:

\[r(a, n) = a^n\]

parameters

`a::Real` : real-valued parameter of the exponential
`n₀::Int` : start  the sample positon signal
`nf::Int` : stop  of the sample positon  signal

returns

s::signal

MySignalProcessing.sigadd — Method

sigadd(s1::signal, s2::signal)

Function to add two signal s1 and s2. pls note that the signals must have the same length If sequences are of unequal lenghts or if the sample positions are different for equal-length sequences, the we cannot directly add We have to first augment x1(n) and x2(n) , so that they have the same position vector n (and hence the same length)

parameters

`s1::signal` : signal to be added
`s2::signal` : signal to be added

returns

`s::signal` : signal s1 + s2

MySignalProcessing.sigfold — Method

sigfold(s::signal)

In this operation, each sample of x(n) is flipped around n=0 to obtain a folded sequence y(n)

y(n) = {x(-n)}

parameters

`s::signal` : signal to be folded

returns

`s::signal` : folded signal

MySignalProcessing.sigmult — Method

sigmult(s1::signal, s2::signal)

Function to multiply two signal s1 and s2. pls note that the signals must have the same length If sequences are of unequal lenghts or if the sample positions are different for equal-length sequences, the we cannot directly add We have to first augment x1(n) and x2(n) , so that they have the same position vector n (and hence the same length)

parameters

`s1::signal` : signal to be added
`s2::signal` : signal to be added

returns

`s::signal` : signal s1 * s2

MySignalProcessing.sigrand — Method

sigrand(start::Int, stop::Int)

Function to generate a random signal starting at sample position start and ending at sample position stop. The function generates a length N random sequence whose elements are uniformely distributed between 0 and 1.

parameters

`start::Int` : start  the sample positon signal
`stop::Int` : stop  of the sample positon  signal

returns

s::signal

MySignalProcessing.sigscale — Method

sigscale(α::Real)

Each signal sample is multiple by a scalar α

parameters

`α::Real` : scalar

returns

anonimous function that takes a signal as argument and returns 
     scaled signal (designed to be composable)

MySignalProcessing.sigshift — Method

sigshift(n₀::Int)

parameters

`n₀::Int` : shift amount

returns

s:: anonimous  function that shifts a signal (designed to be composable)

MySignalProcessing.stepseq — Method

stepseq(start::Int, stop::Int)

Unit Step Sequence This function has the form of :

stepseq(n) = 1 if n ≥ 0
            = 0 if n < 0

parameters

`start:: Int` : start of the sampple position n (horizontal axis)
`stop:: Int` : stop of the sample position n (horizontal axis)

returns

s::signal